requirements.txt
asgiref==3.8.1
backup==0.0.1
blinker==1.8.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
Django==5.0.6
Flask==3.0.3
Flask-JWT-Extended==4.6.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==8.4.0
mysqldump==0.0.10
pycparser==2.22
PyJWT==2.8.0
sqlparse==0.5.0
tzdata==2024.1
Werkzeug==3.0.3
